# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Вектор — planning API",
    "version": "1.0.0",
    "description": "Public scenario, run and geography API. Engine coordinates and money are explicit in ScenarioSpec."
  },
  "servers": [
    {
      "url": "http://localhost:58080"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "ScenarioSpec": {
        "properties": {
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/Zone"
            },
            "title": "Zones",
            "type": "array"
          },
          "sites": {
            "items": {
              "$ref": "#/components/schemas/Site"
            },
            "title": "Sites",
            "type": "array"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/ChargerOption"
            },
            "title": "Options",
            "type": "array"
          },
          "grid_nodes": {
            "items": {
              "$ref": "#/components/schemas/GridNode"
            },
            "title": "Grid Nodes",
            "type": "array"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/Scenario"
            },
            "title": "Scenarios",
            "type": "array"
          },
          "travel_edges": {
            "items": {
              "$ref": "#/components/schemas/TravelEdge"
            },
            "title": "Travel Edges",
            "type": "array"
          },
          "parameters": {
            "$ref": "#/components/schemas/Parameters"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetReference"
            },
            "title": "Datasets",
            "type": "array"
          },
          "locked_site_ids": {
            "items": {
              "type": "string"
            },
            "title": "Locked Site Ids",
            "type": "array"
          },
          "excluded_site_ids": {
            "items": {
              "type": "string"
            },
            "title": "Excluded Site Ids",
            "type": "array"
          }
        },
        "required": [
          "id",
          "zones",
          "sites",
          "options",
          "grid_nodes",
          "scenarios",
          "travel_edges",
          "parameters"
        ],
        "title": "PlanningInput",
        "type": "object"
      },
      "CorridorSpec": {
        "properties": {
          "route_km": {
            "exclusiveMinimum": 0,
            "title": "Route Km",
            "type": "number"
          },
          "battery_usable_kwh": {
            "exclusiveMinimum": 0,
            "title": "Battery Usable Kwh",
            "type": "number"
          },
          "initial_soc": {
            "exclusiveMinimum": 0,
            "maximum": 1,
            "title": "Initial Soc",
            "type": "number"
          },
          "reserve_soc": {
            "exclusiveMaximum": 1,
            "minimum": 0,
            "title": "Reserve Soc",
            "type": "number"
          },
          "consumption_kwh_per_km": {
            "exclusiveMinimum": 0,
            "title": "Consumption Kwh Per Km",
            "type": "number"
          },
          "consumption_multiplier": {
            "default": 1,
            "exclusiveMinimum": 0,
            "title": "Consumption Multiplier",
            "type": "number"
          },
          "stations": {
            "items": {
              "$ref": "#/components/schemas/CorridorStation"
            },
            "title": "Stations",
            "type": "array"
          }
        },
        "required": [
          "route_km",
          "battery_usable_kwh",
          "initial_soc",
          "reserve_soc",
          "consumption_kwh_per_km",
          "stations"
        ],
        "title": "CorridorInput",
        "type": "object"
      },
      "FleetSpec": {
        "properties": {
          "slot_minutes": {
            "default": 15,
            "exclusiveMinimum": 0,
            "maximum": 60,
            "title": "Slot Minutes",
            "type": "integer"
          },
          "horizon_slots": {
            "exclusiveMinimum": 0,
            "maximum": 192,
            "title": "Horizon Slots",
            "type": "integer"
          },
          "efficiency": {
            "default": 0.95,
            "exclusiveMinimum": 0,
            "maximum": 1,
            "title": "Efficiency",
            "type": "number"
          },
          "buses": {
            "items": {
              "$ref": "#/components/schemas/FleetBus"
            },
            "minItems": 1,
            "title": "Buses",
            "type": "array"
          },
          "sites": {
            "items": {
              "$ref": "#/components/schemas/FleetSite"
            },
            "minItems": 1,
            "title": "Sites",
            "type": "array"
          },
          "trips": {
            "items": {
              "$ref": "#/components/schemas/FleetTrip"
            },
            "title": "Trips",
            "type": "array"
          },
          "windows": {
            "items": {
              "$ref": "#/components/schemas/ChargeWindow"
            },
            "title": "Windows",
            "type": "array"
          },
          "solver_seconds": {
            "default": 30,
            "maximum": 3600,
            "minimum": 1,
            "title": "Solver Seconds",
            "type": "integer"
          }
        },
        "required": [
          "horizon_slots",
          "buses",
          "sites",
          "trips",
          "windows"
        ],
        "title": "FleetInput",
        "type": "object"
      },
      "Error": {
        "type": "object",
        "required": [
          "code",
          "detail"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "SavedScenario": {
        "type": "object",
        "required": [
          "id",
          "name",
          "spec"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "spec": {
            "$ref": "#/components/schemas/ScenarioSpec"
          },
          "sha256": {
            "type": "string"
          }
        }
      },
      "ScenarioSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "sha256",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Run": {
        "type": "object",
        "required": [
          "id",
          "state"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "scenario_id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ]
          },
          "attempts": {
            "type": "integer"
          },
          "error_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "error_detail": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PlanResult": {
        "type": "object",
        "required": [
          "optimization",
          "simulation",
          "explanations",
          "metadata"
        ],
        "properties": {
          "optimization": {
            "$ref": "#/components/schemas/OptimizationResult"
          },
          "simulation": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SimulationResult"
            }
          },
          "explanations": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "metadata": {
            "type": "object"
          }
        }
      },
      "DatasetManifest": {
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "source",
          "checksum",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "observed",
              "derived",
              "assumed"
            ]
          },
          "source": {
            "type": "string"
          },
          "license": {
            "type": [
              "string",
              "null"
            ]
          },
          "checksum": {
            "type": "string"
          },
          "captured_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/scenarios": {
      "get": {
        "summary": "List saved scenario snapshots",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScenarioSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Validate and create immutable scenario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "spec"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "spec": {
                    "$ref": "#/components/schemas/ScenarioSpec"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedScenario"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scenarios/{id}": {
      "get": {
        "summary": "Get scenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedScenario"
                }
              }
            }
          },
          "404": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scenarios/{id}/runs": {
      "post": {
        "summary": "Queue idempotent calculation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "404": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{id}": {
      "get": {
        "summary": "Get run state",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "404": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{id}/cancel": {
      "post": {
        "summary": "Cancel queued or running run",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{id}/results": {
      "get": {
        "summary": "Get completed plan and simulation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResult"
                }
              }
            }
          },
          "404": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{id}/events": {
      "get": {
        "summary": "SSE stream with Last-Event-ID replay",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Run events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets": {
      "get": {
        "summary": "List imported dataset versions",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatasetManifest"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/map": {
      "get": {
        "summary": "GeoJSON in bounding box",
        "parameters": [
          {
            "name": "bbox",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "west,south,east,north in WGS84"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "FeatureCollection",
            "content": {
              "application/geo+json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tiles/{z}/{x}/{y}": {
      "get": {
        "summary": "PostGIS vector tile",
        "parameters": [
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mapbox Vector Tile",
            "content": {
              "application/vnd.mapbox-vector-tile": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/corridors/check": {
      "post": {
        "summary": "Check route reachability and station outages",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorridorSpec"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fleets/schedule": {
      "post": {
        "summary": "Schedule charging for an assigned vehicle fleet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FleetSpec"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}